Try fix pid lock missing issue

+++ fluxmonitor/launcher.py
@@ -103,6 +103,7 @@ def init_service(klass_name, options):
 def deamon_entry(options, service=None):
     pid_handler = None
     server = None
+    pid = os.getpid()
 
     # Close all file descriptor except stdin/stdout/stderr and pid file
     # descriptor
@@ -123,7 +124,7 @@ def deamon_entry(options, service=None):
 
                 try:
                     pid_handler = lock_pidfile(options)
-                    pid_handler.write(repr(os.getpid()))
+                    pid_handler.write(repr(pid))
                     pid_handler.flush()
 
                     sys.stdin.close()
@@ -181,7 +182,7 @@ def deamon_entry(options, service=None):
     else:
         try:
             pid_handler = lock_pidfile(options)
-            pid_handler.write(repr(os.getpid()))
+            pid_handler.write(repr(pid))
         except FatalException as e:
             return e.args[0]
 
@@ -207,10 +208,10 @@ def deamon_entry(options, service=None):
             return 1
 
     finally:
-        fcntl.lockf(pid_handler.fileno(), fcntl.LOCK_UN)
-        pid_handler.close()
-        os.unlink(options.pidfile)
-
+        if os.getpid() == pid:
+            fcntl.lockf(pid_handler.fileno(), fcntl.LOCK_UN)
+            pid_handler.close()
+            os.unlink(options.pidfile)
 
     return 0
